fixed scripts
