+++ app/src/mocks/challenges.ts
@@ -3,10 +3,7 @@ import { ChallengePayload } from 'types/api';
 let i = 100;
 
 // TODO: refactor for types, collection, interfaces etc
-export const mockChallenges: Omit<
-    ChallengePayload,
-    'pubkey' | 'eventPubkey' | 'eventId'
->[] = [
+export const mockChallenges: Omit<ChallengePayload, 'pubkey'>[] = [
     // Heavy Duty Social
     {
         key: ++i,
